+++ provisioning/site-production/01-remote-state.tf
@@ -9,8 +9,8 @@
 # The BAS-AWS remote state includes outputs such as identifiers for VPC subnets, VPC security groups and DNS zones
 #
 # State source: https://s3-eu-west-1.amazonaws.com/bas-terraform-remote-state-prod/v1/BAS-AWS/terraform.tfstate
-# Terraform source: https://www.terraform.io/docs/providers/terraform/r/remote_state.html
-resource "terraform_remote_state" "BAS-AWS" {
+# Terraform source: https://www.terraform.io/docs/providers/terraform/d/remote_state.html
+data "terraform_remote_state" "BAS-AWS" {
     backend = "s3"
     config {
         bucket = "bas-terraform-remote-state-prod"
@@ -22,8 +22,8 @@ resource "terraform_remote_state" "BAS-AWS" {
 # The BAS-PACKER-VM-TEMPLATES remote state includes AMI identifiers for EC2 instances
 #
 # State source: https://s3-eu-west-1.amazonaws.com/bas-terraform-remote-state-dev/v1/BAS-PACKER-VM-TEMPLATES/terraform.tfstate
-# Terraform source: https://www.terraform.io/docs/providers/terraform/r/remote_state.html
-resource "terraform_remote_state" "BAS-PACKER-VM-TEMPLATES" {
+# Terraform source: https://www.terraform.io/docs/providers/terraform/d/remote_state.html
+data "terraform_remote_state" "BAS-PACKER-VM-TEMPLATES" {
     backend = "s3"
     config {
         bucket = "bas-terraform-remote-state-prod"